+++ openfisca_survey_manager/policy/tests/test_compute_winners_losers.py
@@ -1,7 +1,7 @@
 import pytest
 from openfisca_country_template.reforms.modify_social_security_taxation import modify_social_security_taxation
 
-from openfisca_survey_manager.simulations import SecretViolationError
+from openfisca_survey_manager.policy import SecretViolationError
 from openfisca_survey_manager.tests.test_scenario import create_randomly_initialized_survey_scenario